+++ src/bootstrap.js
@@ -1,23 +1,48 @@
 const fs = require('fs');
-const path = require('path');
 const vscode = require('vscode');
 const axios = require('axios');
 
-let statusBarItem = null;
-let cachedClasses = null;
-let cachedFileMtime = null;
+const extractCssClasses = (css) => {
+  try {
+    const classRegex = /\.([a-zA-Z0-9\-_]+)([^{]*?)\s*{([^}]*)}/gs;
+    const classes = [];
+    let match;
+
+    while ((match = classRegex.exec(css))) {
+      const className = match[1];
+      let classProperties = match[0];
+
+      classProperties = classProperties
+        .replace(/\s*{\s*/, ' {\n  ')
+        .replace(/;\s*/g, ';\n  ')
+        .replace(/\s*}\s*$/, '\n}');
+
+      classes.push({
+        className: className,
+        classProperties: classProperties,
+      });
+    }
+
+    return classes;
+  } catch (error) {
+    console.error(`Error extracting CSS classes: ${error}`);
+    return [];
+  }
+};
 
 const getBootstrapCdnLink = async () => {
+  const cdnRegex = /<link[^>]+href=["']((?!cdn)[^"']*bootstrap[^"']*\.css)["']/;
+
   try {
     const htmlFiles = await vscode.workspace.findFiles('**/*.html');
-    const cdnRegex =
-      /<link[^>]+href=["'](https:\/\/cdn\.jsdelivr\.net\/npm\/bootstrap@[\d\.]+\/dist\/css\/bootstrap\.min\.css)["']/;
 
     for (const file of htmlFiles) {
-      const content = fs.readFileSync(file.fsPath, 'utf8');
-      const match = cdnRegex.exec(content);
+      const content = await fs.promises.readFile(file.fsPath, 'utf8');
+      const match = content.match(cdnRegex);
+
       if (match) {
-        return match[1];
+        const cdnLink = match[1];
+        return cdnLink;
       }
     }
 
@@ -28,75 +53,15 @@ const getBootstrapCdnLink = async () => {
   }
 };
 
-const getBootstrapVersion = (cdnLink) => {
-  const versionRegex = /@(\d+\.\d+\.\d+)/;
-  const matches = cdnLink.match(versionRegex);
-  if (matches && matches[1]) {
-    return matches[1];
-  }
-};
-
-const setStatusBarItem = async () => {
-  try {
-    const packageJsonPath = await vscode.workspace.findFiles(
-      '**/node_modules/bootstrap/package.json',
-    );
-    let bootstrapVersion;
-
-    if (packageJsonPath[0]) {
-      const packageJson = JSON.parse(fs.readFileSync(path.join(packageJsonPath[0].fsPath), 'utf8'));
-      bootstrapVersion = packageJson.version;
-    } else {
-      bootstrapVersion = getBootstrapVersion(await getBootstrapCdnLink());
-    }
-
-    if (!statusBarItem) {
-      statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
-    }
-
-    if (bootstrapVersion) {
-      statusBarItem.text = `$(bootstrap-icon) Bootstrap v${bootstrapVersion
-        .split('.')
-        .slice(0, 2)
-        .join('.')}`;
-      statusBarItem.tooltip = 'Your current Bootstrap version in this project';
-      statusBarItem.show();
-    }
-  } catch (error) {
-    return null;
-  }
-};
-
-const extractCssClasses = (css) => {
-  try {
-    const classRegex = /(?:^|\s)\.([\w-]+)\s*{([^}]*)}/gs;
-    const classes = [];
-    let match;
-
-    while ((match = classRegex.exec(css))) {
-      const className = match[1];
-      let classContent = match[2].trim();
-
-      classContent = '{\n  ' + classContent;
-      classContent = classContent.replace(/;\s*/g, ';\n  ');
-      classContent = classContent.replace(/\s*$/, '\n}');
-
-      classes.push({
-        className,
-        classContent: `.${className} ${classContent}`,
-      });
-    }
-
-    return classes;
-  } catch (error) {
-    vscode.window.showWarningMessage(`Error extracting CSS classes: ${error}`);
-    return [];
-  }
-};
-
 const getCssFromCdn = async (url) => {
   try {
-    const bootstrapVersion = getBootstrapVersion(url);
+    let bootstrapVersion = null;
+    const versionRegex = /@(\d+\.\d+\.\d+)/;
+    const matches = url.match(versionRegex);
+    if (matches && matches[1]) {
+      bootstrapVersion = matches[1];
+    }
+
     const response = await axios.get(
       `https://cdn.jsdelivr.net/npm/bootstrap@${bootstrapVersion}/dist/css/bootstrap.css`,
     );
@@ -107,61 +72,68 @@ const getCssFromCdn = async (url) => {
   }
 };
 
-const getBsClasses = async () => {
+const getCssFromLocalFiles = async () => {
+  const htmlFiles = await vscode.workspace.findFiles('**/*.html');
+  const bootstrapLinkRegex = /<link[^>]+href=["']((?!.*jsdelivr).*\bbootstrap\b[^"']*)["']/g;
+  let searchPatterns = ['**/node_modules/bootstrap/dist/css/bootstrap.css'];
+
+  for (const file of htmlFiles) {
+    const fileContent = await fs.promises.readFile(file.fsPath, 'utf8');
+    const bootstrapLinkMatches = [...fileContent.matchAll(bootstrapLinkRegex)];
+
+    for (const match of bootstrapLinkMatches) {
+      const bootstrapFilePath = match[1];
+      searchPatterns.push(bootstrapFilePath.split('./')[1]);
+    }
+  }
+
+  try {
+    for (const pattern of searchPatterns) {
+      const bootstrapFiles = await vscode.workspace.findFiles(pattern);
+      for (const file of bootstrapFiles) {
+        const bootstrapPath = file.fsPath;
+        if (fs.existsSync(bootstrapPath)) {
+          const css = await fs.promises.readFile(bootstrapPath, 'utf8');
+          return css;
+        } else {
+          vscode.window.showInformationMessage(`Bootstrap file not found: ${bootstrapPath}`);
+        }
+      }
+    }
+    return null;
+  } catch (error) {
+    vscode.window.showInformationMessage(`Error finding Bootstrap file in local files: ${error}`);
+    return null;
+  }
+};
+
+const getBootstrapClasses = async () => {
   try {
     const workspaceFolders = vscode.workspace.workspaceFolders;
     if (!workspaceFolders || workspaceFolders.length === 0) {
       return null; // No workspace is open
     }
 
-    const bootstrapFiles = await vscode.workspace.findFiles(
-      '**/node_modules/bootstrap/dist/css/bootstrap.css',
-    );
+    let css = await getCssFromLocalFiles();
 
-    if (bootstrapFiles.length > 0) {
-      const bootstrapPath = bootstrapFiles[0].fsPath;
-
-      try {
-        if (fs.existsSync(bootstrapPath)) {
-          const stats = fs.statSync(bootstrapPath);
-          const mtime = stats.mtime.getTime();
-
-          // If the file has not changed, return the cached value
-          if (cachedFileMtime === mtime) {
-            return cachedClasses;
-          }
-
-          // File has changed or is being loaded for the first time
-          const css = fs.readFileSync(bootstrapPath, 'utf8');
-          cachedClasses = extractCssClasses(css);
-          cachedFileMtime = mtime;
-
-          return cachedClasses;
-        } else {
-          vscode.window.showInformationMessage(`Bootstrap file not found: ${bootstrapPath}`);
-          return null;
-        }
-      } catch (error) {
-        vscode.window.showInformationMessage(`Error reading Bootstrap file: ${error}`);
-        return null;
-      }
-    } else {
+    if (!css) {
       const cdnLink = await getBootstrapCdnLink();
       if (cdnLink) {
-        const css = await getCssFromCdn(cdnLink);
-        if (css) {
-          cachedClasses = extractCssClasses(css);
-          return cachedClasses;
-        }
+        css = await getCssFromCdn(cdnLink);
       }
     }
+
+    if (css) {
+      return extractCssClasses(css);
+    }
+
     return null;
   } catch (error) {
+    vscode.window.showInformationMessage(`Error finding Bootstrap classes: ${error}`);
     return null;
   }
 };
 
 module.exports = {
-  getBsClasses,
-  setStatusBarItem,
+  getBootstrapClasses,
 };